Create ds-project-7-ford-vs-ferrari-by-ivanbaukin.ipynb
Car classification project based on Xception model, which was pretrained on ImageNet. In project we compare different ways to use basic parametres such as epochs and learning rate and saw how it effect on score. Also we used fine-tuning with differents freeze (differents unfreeze layers in basemodel). We had a notebook where we used EfficientNetB6 pretrained on ImageNet ( transfer learning) like a basemodel, but it gave worse score than Efficient.
But i think this problem linked with wrong tuning basemodel.
So, finally we have result 0,94 accuracy, and I had some idea how to improve this score. But today is deadline of this project, thats why i attach the current notebook.